+++ src/pages/Accessories/AccessoriesPage.tsx
@@ -1,5 +1,10 @@
+import { Catalog } from '../../components/Catalog/Catalog';
 import '../../pages/pages.scss';
 
 export const AccessoriesPage = () => {
-  return <div className="page__container">AccessoriesPage</div>;
+  return (
+    <div className="page__container">
+      <Catalog productType="Accessories" />
+    </div>
+  );
 };
