+++ src/pages/Tablets/TabletsPage.tsx
@@ -1,3 +1,9 @@
+import { Catalog } from '../../components/Catalog/Catalog';
+
 export const TabletsPage = () => {
-  return <div className="page__container">TabletsPage</div>;
+  return (
+    <div className="page__container">
+      <Catalog productType="Tablets" />
+    </div>
+  );
 };
